wip: client tools installer & builds
At this point, there's a rough portable shell script, with a few FIXMEs in it.
The logic works to install the "stable" release, if explicitly pointed at a
local file.

No CDN dependencies are in the script.  We need to get the builds going and the
metadata into the CDN to update the URLs and reference those.  For the time
being, I'm pointing the stables directly at GitHub's artifacts area.