Cloudflare process docs; nightlies stub script
Documented a bunch of the web UI configuration done with CloudFlare.

Have a script `build-nightlies.sh` which so far does exactly that, and checks
that it has valid credentials to publish, but doesn't yet publish.  This
provides a basis for working on the publishing side, so committing at this
stage.