upstream/pdp-personal: _nats support extra-verbose
The zsh tab-completion for `nats` defaults to providing the names of contexts,
but not their descriptions.  Getting the description is a jq invocation per
file, which is comparatively heavy-weight and slow enough to be noticeable
given as many contexts as I have.

But zsh has the `extra-verbose` zstyle for just this situation.  So use it.
Provide copious comments showing exactly how to use it.

Leaves a TODO to figure out caching optimizations.